crud salvo update para detalle y entrada

+++ inventario/src/main/java/com/misiontic2022/inventario/repository/EntradasRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.misiontic2022.inventario.domain.Entrada;
 
-public interface EntradasRepository extends JpaRepository<Entrada, Integer> {
+public interface EntradasRepository extends JpaRepository<Entrada, Long> {
+
+    // void update(Entrada entrada);
 
 }
